src/main/java/demo/exercise/Exercise.java
package demo.exercise;

import org.springframework.data.annotation.Id;

public class Exercise {

    @Id
    private String id;

    private String exerciseName;
    private String muscleGroup;
    private int reps;


    public String getExerciseName() {
        return exerciseName;
    }

    public void setExerciseName(String exerciseName) {
        this.exerciseName = exerciseName;
    }

    public String getMuscleGroup() {
        return muscleGroup;
    }

    public void setMuscleGroup(String muscleGroup) {
        this.muscleGroup = muscleGroup;
    }

    public int getReps() {
        return reps;
    }

    public void setReps(int reps) {
        this.reps = reps;
    }


}

